feat: add check if the author exists

+++ src/application/use-cases/remove-author.ts
@@ -1,11 +1,14 @@
-import { type DeleteAuthor } from '@/application/contracts/repositories'
+import { type GetAuthor, type DeleteAuthor } from '@/application/contracts/repositories'
+import { AuthorNotFoundError } from '../errors'
 
 export class RemoveAuthorUseCase {
   constructor(
-    private readonly authorRepository: DeleteAuthor
+    private readonly authorRepository: DeleteAuthor & GetAuthor
   ) { }
 
   async execute({ authorId }: Input): Promise<void> {
+    const author = await this.authorRepository.get({ id: authorId })
+    if (author === undefined) throw new AuthorNotFoundError()
     await this.authorRepository.delete({ id: authorId })
   }
 }
